LayoutTests/imported/w3c/web-platform-tests/workers/modules/dedicated-worker-import-data-url-cross-origin-expected.txt

CONSOLE MESSAGE: Origin null is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: Importing a module script failed.

Harness Error (FAIL), message = Script error.

PASS static import data url from data: URL should be allowed.
TIMEOUT static import script from data: URL should be allowed. Test timed out
NOTRUN dynamic import data url from data: URL should be allowed.
NOTRUN dynamic import script from data: URL should be blocked.
